requirements.txt
flask
flask-cors
python-docx
pydantic
python-multipart
reportlab
PyPDF2
openai
pdfkit
jinja2
python-dotenv
fpdf
PyMuPDF>=1.24.9
celery
redis
dirtyjson
json-repair
pyyaml
pytest-env
jsonschema
playwright==1.47.0
croniter
rapidfuzz>=3.14.0,<3.15
environs
